Move downloadFile to tools/common, apply exclusive mkdir lock in dev script
