+++ scripts/download-php-binary.ts
@@ -21,11 +21,11 @@ import crypto from 'crypto';
 import fs from 'fs';
 import os from 'os';
 import path from 'path';
-import { Readable } from 'stream';
-import { pipeline } from 'stream/promises';
 import { extract } from 'tar';
 import { z } from 'zod';
+import { downloadFile } from '../tools/common/lib/download-file';
 import { extractZip } from '../tools/common/lib/extract-zip';
+import { isErrnoException } from '../tools/common/lib/is-errno-exception';
 import {
 	buildPhpBinaryUrl,
 	getPhpBinaryHash,
@@ -34,7 +34,6 @@ import {
 } from '../tools/common/lib/php-binary-metadata';
 import { getConfigDirectory } from '../tools/common/lib/well-known-paths';
 import { RecommendedPHPVersion, SupportedPHPVersions } from '../tools/common/types/php-versions';
-import type { ReadableStream as NodeReadableStream } from 'stream/web';
 
 const versionSchema = z.enum( SupportedPHPVersions );
 const platformSchema = z.enum( [ 'darwin', 'win32', 'linux' ] );
@@ -80,28 +79,21 @@ async function computeAndPrintHash(): Promise< void > {
 	}
 
 	const url = buildPhpBinaryUrl( version, args.platform, args.arch );
-	const filename = path.basename( url );
-	const tmpPath = path.join( os.tmpdir(), filename );
+	const tmpPath = path.join( os.tmpdir(), `${ process.pid }-${ path.basename( url ) }` );
 
-	console.log( `Downloading ${ filename } to compute hash…` );
-	const response = await fetch( url );
-	if ( ! response.ok ) {
-		throw new Error( `Download failed: HTTP ${ response.status }` );
+	console.log( `Downloading ${ path.basename( url ) } to compute hash…` );
+	try {
+		await downloadFile( url, tmpPath );
+		const data = await fs.promises.readFile( tmpPath );
+		const hash = crypto.createHash( 'sha256' ).update( data ).digest( 'hex' );
+		const key = `'${ version }-${ args.platform }-${ effectiveArch }'`;
+		console.log( `${ key }: '${ hash }',` );
+		console.log(
+			`\nAdd this entry to PHP_BINARY_HASHES in tools/common/lib/php-binary-metadata.ts`
+		);
+	} finally {
+		fs.rmSync( tmpPath, { force: true } );
 	}
-	await pipeline(
-		Readable.fromWeb( response.body! as NodeReadableStream< Uint8Array > ),
-		fs.createWriteStream( tmpPath )
-	);
-
-	const data = await fs.promises.readFile( tmpPath );
-	const hash = crypto.createHash( 'sha256' ).update( data ).digest( 'hex' );
-	fs.unlinkSync( tmpPath );
-
-	const key = `'${ version }-${ args.platform }-${ effectiveArch }'`;
-	console.log( `${ key }: '${ hash }',` );
-	console.log(
-		`\nAdd this entry to PHP_BINARY_HASHES in apps/cli/lib/dependency-management/php-binary-metadata.ts`
-	);
 }
 
 async function main(): Promise< void > {
@@ -115,12 +107,9 @@ async function main(): Promise< void > {
 
 	const isWindows = args.platform === 'win32';
 	const url = buildPhpBinaryUrl( version, args.platform, args.arch );
-	const filename = path.basename( url );
 	const binDir = path.join( getConfigDirectory(), 'php-bin', version );
 	const binaryName = isWindows ? 'php.exe' : 'php';
 	const destPath = path.join( binDir, binaryName );
-	const tmpDir = os.tmpdir();
-	const downloadPath = path.join( tmpDir, filename );
 	const platformKey = `${ args.platform }-${ effectiveArch }`;
 
 	try {
@@ -131,72 +120,89 @@ async function main(): Promise< void > {
 			return;
 		}
 
-		fs.mkdirSync( binDir, { recursive: true } );
-
-		console.log( `Downloading PHP ${ version } (${ patchVersion }) for ${ platformKey }…` );
-		console.log( `  URL: ${ url }` );
-
-		const response = await fetch( url );
-		if ( ! response.ok ) {
-			throw new Error( `Download failed: HTTP ${ response.status } ${ response.statusText }` );
-		}
-		await pipeline(
-			Readable.fromWeb( response.body! as NodeReadableStream< Uint8Array > ),
-			fs.createWriteStream( downloadPath )
-		);
-		console.log( 'Download complete.' );
-
-		// Verify SHA-256
-		const expected = getPhpBinaryHash( version, args.platform, args.arch );
-		if ( expected ) {
-			console.log( 'Verifying SHA-256…' );
-			const data = await fs.promises.readFile( downloadPath );
-			const actual = crypto.createHash( 'sha256' ).update( data ).digest( 'hex' );
-			if ( actual !== expected ) {
-				throw new Error( `SHA-256 mismatch:\n  expected ${ expected }\n  got      ${ actual }` );
+		// Ensure ~/.studio/php-bin/ exists, then atomically claim this version's slot.
+		fs.mkdirSync( path.dirname( binDir ), { recursive: true } );
+		try {
+			fs.mkdirSync( binDir );
+		} catch ( err ) {
+			if ( isErrnoException( err ) && err.code === 'EEXIST' ) {
+				console.log(
+					`PHP ${ version } is already being downloaded by another process. Waiting…`
+				);
+				return;
 			}
-			console.log( '  Hash OK.' );
-		} else {
-			console.warn(
-				`Warning: no pinned hash for ${ version }-${ platformKey }. Skipping verification.\n` +
-					`         Compute it with: npx ts-node scripts/download-php-binary.ts --compute-hashes ${ version } ${ args.platform } ${ args.arch }`
+			throw err;
+		}
+
+		const downloadPath = path.join( binDir, path.basename( url ) );
+
+		try {
+			console.log( `Downloading PHP ${ version } (${ patchVersion }) for ${ platformKey }…` );
+			console.log( `  URL: ${ url }` );
+			await downloadFile( url, downloadPath, ( downloaded, total ) => {
+				const dl = ( downloaded / 1024 / 1024 ).toFixed( 1 );
+				const tot = total ? ` / ${ ( total / 1024 / 1024 ).toFixed( 1 ) } MB` : '';
+				process.stdout.write( `\r  ${ dl } MB${ tot }` );
+			} );
+			console.log( '\nDownload complete.' );
+
+			// Verify SHA-256
+			const expected = getPhpBinaryHash( version, args.platform, args.arch );
+			if ( expected ) {
+				console.log( 'Verifying SHA-256…' );
+				const data = await fs.promises.readFile( downloadPath );
+				const actual = crypto.createHash( 'sha256' ).update( data ).digest( 'hex' );
+				if ( actual !== expected ) {
+					throw new Error( `SHA-256 mismatch:\n  expected ${ expected }\n  got      ${ actual }` );
+				}
+				console.log( '  Hash OK.' );
+			} else {
+				console.warn(
+					`Warning: no pinned hash for ${ version }-${ platformKey }. Skipping verification.\n` +
+						`         Compute it with: npx ts-node scripts/download-php-binary.ts --compute-hashes ${ version } ${ args.platform } ${ args.arch }`
+				);
+			}
+
+			// Extract
+			console.log( 'Extracting PHP binary…' );
+			const tmpDir = os.tmpdir();
+			if ( isWindows ) {
+				await extractZip( downloadPath, tmpDir );
+				const src = path.join( tmpDir, 'php.exe' );
+				if ( ! fs.existsSync( src ) ) {
+					throw new Error( `php.exe not found after extraction.` );
+				}
+				fs.copyFileSync( src, destPath );
+				fs.unlinkSync( src );
+			} else {
+				const extractDir = path.join(
+					tmpDir,
+					`php-${ patchVersion }-${ args.platform }-${ effectiveArch }`
+				);
+				fs.mkdirSync( extractDir, { recursive: true } );
+				try {
+					await extract( { file: downloadPath, cwd: extractDir } );
+					const src = path.join( extractDir, 'php' );
+					if ( ! fs.existsSync( src ) ) {
+						throw new Error( `php binary not found after extraction.` );
+					}
+					fs.copyFileSync( src, destPath );
+					fs.chmodSync( destPath, 0o755 );
+				} finally {
+					fs.rmSync( extractDir, { recursive: true, force: true } );
+				}
+			}
+
+			const stats = fs.statSync( destPath );
+			console.log(
+				`\nPHP ${ version } binary installed: ${ destPath } (${ ( stats.size / 1024 / 1024 ).toFixed( 1 ) } MB)`
 			);
+		} catch ( err ) {
+			fs.rmSync( binDir, { recursive: true, force: true } );
+			throw err;
+		} finally {
+			fs.rmSync( downloadPath, { force: true } );
 		}
-
-		// Extract
-		console.log( 'Extracting PHP binary…' );
-		if ( isWindows ) {
-			await extractZip( downloadPath, tmpDir );
-			const src = path.join( tmpDir, 'php.exe' );
-			if ( ! fs.existsSync( src ) ) {
-				throw new Error( `php.exe not found after extraction.` );
-			}
-			fs.copyFileSync( src, destPath );
-			fs.unlinkSync( src );
-		} else {
-			const extractDir = path.join(
-				tmpDir,
-				`php-${ patchVersion }-${ args.platform }-${ effectiveArch }`
-			);
-			fs.mkdirSync( extractDir, { recursive: true } );
-			await extract( { file: downloadPath, cwd: extractDir } );
-			const src = path.join( extractDir, 'php' );
-			if ( ! fs.existsSync( src ) ) {
-				throw new Error( `php binary not found after extraction.` );
-			}
-			fs.copyFileSync( src, destPath );
-			fs.chmodSync( destPath, 0o755 );
-			fs.rmSync( extractDir, { recursive: true, force: true } );
-		}
-
-		fs.unlinkSync( downloadPath );
-
-		const stats = fs.statSync( destPath );
-		console.log(
-			`\nPHP ${ version } binary installed: ${ destPath } (${ ( stats.size / 1024 / 1024 ).toFixed(
-				1
-			) } MB)`
-		);
 	} catch ( error ) {
 		console.warn( `Warning: PHP binary download failed — ${ ( error as Error ).message }` );
 		console.warn(
